Extract claude_hook.rs and install.rs (refs #86)
Moves hook event handling (handle_claude_hook + describe_tool_use) into
claude_hook.rs (166 LoC) and hook/shell installation helpers
(install_claude_hooks, cleanup_legacy_claude_hooks, uninstall_claude_hooks,
install_shell_integration) into install.rs (122 LoC).

main.rs drops from 1,080 to 799 LoC.

No behavior change — pure mechanical refactor.
